+++ release-notes.md
@@ -13,3 +13,10 @@
 
 + added support to create custom roles for specific user behaviour
 + updated check of all entities operations based on new roles
+
+## Version 0.6.1
+
+2022.11.18
+
++ added range attribute for metadata and topic fields in experiments \[min,max\] for scalar
++ added enum type for experiments metadata and topic fields between values compiled in "range" attribute in protocol